refactor(common): #345 - Extract RelativeCommonMap to separate api package
