+++ pkg/controller/common/customize/manager.go
@@ -18,6 +18,7 @@ package customize
 
 import (
 	"fmt"
+	commonv1 "metacontroller/pkg/controller/common/api/v1"
 	v1 "metacontroller/pkg/controller/common/customize/api/v1"
 	"metacontroller/pkg/hooks"
 
@@ -333,8 +334,8 @@ func listObjects(selector labels.Selector, namespace string, informer *dynamicin
 	return informer.Lister().List(selector)
 }
 
-func (rm *Manager) GetRelatedObjects(parent *unstructured.Unstructured) (common.RelativeObjectMap, error) {
-	childMap := make(common.RelativeObjectMap)
+func (rm *Manager) GetRelatedObjects(parent *unstructured.Unstructured) (commonv1.RelativeObjectMap, error) {
+	childMap := make(commonv1.RelativeObjectMap)
 	if !rm.IsEnabled() {
 		return childMap, nil
 	}
