+++ pkg/controller/composite/controller.go
@@ -19,6 +19,7 @@ package composite
 import (
 	"context"
 	"fmt"
+	commonv1 "metacontroller/pkg/controller/common/api/v1"
 	"metacontroller/pkg/hooks"
 	"metacontroller/pkg/logging"
 	"reflect"
@@ -530,7 +531,7 @@ func (pc *parentController) syncParentObject(parent *unstructured.Unstructured) 
 	if err != nil {
 		return err
 	}
-	desiredChildren := common.MakeRelativeObjectMap(parent, syncResult.Children)
+	desiredChildren := commonv1.MakeRelativeObjectMap(parent, syncResult.Children)
 
 	// Enqueue a delayed resync, if requested.
 	if syncResult.ResyncAfterSeconds > 0 {
@@ -658,7 +659,7 @@ func (pc *parentController) canAdoptFunc(parent *unstructured.Unstructured) func
 	})
 }
 
-func (pc *parentController) claimChildren(parent *unstructured.Unstructured) (common.RelativeObjectMap, error) {
+func (pc *parentController) claimChildren(parent *unstructured.Unstructured) (commonv1.RelativeObjectMap, error) {
 	// Set up values common to all child types.
 	parentNamespace := parent.GetNamespace()
 	parentGVK := pc.parentResource.GroupVersionKind()
@@ -669,7 +670,7 @@ func (pc *parentController) claimChildren(parent *unstructured.Unstructured) (co
 	canAdoptFunc := pc.canAdoptFunc(parent)
 
 	// Claim all child types.
-	childMap := make(common.RelativeObjectMap)
+	childMap := make(commonv1.RelativeObjectMap)
 	for _, child := range pc.cc.Spec.ChildResources {
 		// List all objects of the child kind in the parent object's namespace,
 		// or in all namespaces if the parent is cluster-scoped.
