+++ pkg/controller/decorator/controller.go
@@ -19,6 +19,7 @@ package decorator
 import (
 	"context"
 	"fmt"
+	commonv1 "metacontroller/pkg/controller/common/api/v1"
 	v1 "metacontroller/pkg/controller/decorator/api/v1"
 	"metacontroller/pkg/hooks"
 	"reflect"
@@ -549,7 +550,7 @@ func (c *decoratorController) syncParentObject(parent *unstructured.Unstructured
 	if err != nil {
 		return err
 	}
-	desiredChildren := common.MakeRelativeObjectMap(parent, syncResult.Attachments)
+	desiredChildren := commonv1.MakeRelativeObjectMap(parent, syncResult.Attachments)
 
 	// Enqueue a delayed resync, if requested.
 	if syncResult.ResyncAfterSeconds > 0 {
@@ -663,10 +664,10 @@ func (c *decoratorController) syncParentObject(parent *unstructured.Unstructured
 	return manageErr
 }
 
-func (c *decoratorController) getChildren(parent *unstructured.Unstructured) (common.RelativeObjectMap, error) {
+func (c *decoratorController) getChildren(parent *unstructured.Unstructured) (commonv1.RelativeObjectMap, error) {
 	parentUID := parent.GetUID()
 	parentNamespace := parent.GetNamespace()
-	childMap := make(common.RelativeObjectMap)
+	childMap := make(commonv1.RelativeObjectMap)
 
 	for _, child := range c.dc.Spec.Attachments {
 		// List all objects of the child kind in the parent object's namespace,
